[refact] Wrap updates to an `OfflineIndex` inside a transaction
Write operations on an `OfflineIndex` must now be wrapped inside a transaction. The code will assert/crash if that’s not the case.

+++ Tests/OfflineClientTests.swift
@@ -1,9 +1,24 @@
 //
-//  OfflineIndexTests.swift
-//  AlgoliaSearch
+//  Copyright (c) 2016 Algolia
+//  http://www.algolia.com/
 //
-//  Created by Clément Le Provost on 26/08/16.
-//  Copyright © 2016 Algolia. All rights reserved.
+//  Permission is hereby granted, free of charge, to any person obtaining a copy
+//  of this software and associated documentation files (the "Software"), to deal
+//  in the Software without restriction, including without limitation the rights
+//  to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+//  copies of the Software, and to permit persons to whom the Software is
+//  furnished to do so, subject to the following conditions:
+//
+//  The above copyright notice and this permission notice shall be included in
+//  all copies or substantial portions of the Software.
+//
+//  THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+//  IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+//  FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+//  AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+//  LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+//  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+//  THE SOFTWARE.
 //
 
 import AlgoliaSearch
@@ -45,21 +60,25 @@ class OfflineClientTests: OfflineTestCase {
                 // Check that the index does not exist yet.
                 assert(name != index.name)
             }
+            index.beginTransaction()
             index.addObject(self.objects["snoopy"]!) { (content, error) in
                 assert(error == nil)
-                self.client.listIndexesOffline { (content, error) in
-                    guard let content = content else { assert(false); return }
-                    guard let items = content["items"] as? [[String: AnyObject]] else { assert(false); return }
-                    var found = false
-                    for item in items {
-                        guard let name = item["name"] as? String else { assert(false); return }
-                        // Check that the index *does* exist.
-                        if name == index.name {
-                            found = true
+                index.commitTransaction() { (content, error) in
+                    guard error == nil else { assert(false); return }
+                    self.client.listIndexesOffline { (content, error) in
+                        guard let content = content else { assert(false); return }
+                        guard let items = content["items"] as? [[String: AnyObject]] else { assert(false); return }
+                        var found = false
+                        for item in items {
+                            guard let name = item["name"] as? String else { assert(false); return }
+                            // Check that the index *does* exist.
+                            if name == index.name {
+                                found = true
+                            }
                         }
+                        assert(found)
+                        expectation.fulfill()
                     }
-                    assert(found)
-                    expectation.fulfill()
                 }
             }
         }
@@ -69,14 +88,18 @@ class OfflineClientTests: OfflineTestCase {
     func testDeleteIndex() {
         let expectation = expectationWithDescription(#function)
         let index = client.getOfflineIndex(#function)
+        index.beginTransaction()
         index.addObjects(Array(objects.values)) { (content, error) in
             assert(error == nil)
-            assert(self.client.hasOfflineData(index.name))
-            self.client.deleteIndexOffline(index.name) { (content, error) in
-                guard let content = content else { assert(false); return }
-                assert(content["deletedAt"] as? String != nil)
-                assert(!self.client.hasOfflineData(index.name))
-                expectation.fulfill()
+            index.commitTransaction() { (content, error) in
+                guard error == nil else { assert(false); return }
+                assert(self.client.hasOfflineData(index.name))
+                self.client.deleteIndexOffline(index.name) { (content, error) in
+                    guard let content = content else { assert(false); return }
+                    assert(content["deletedAt"] as? String != nil)
+                    assert(!self.client.hasOfflineData(index.name))
+                    expectation.fulfill()
+                }
             }
         }
         waitForExpectationsWithTimeout(expectationTimeout, handler: nil)
@@ -86,20 +109,24 @@ class OfflineClientTests: OfflineTestCase {
         let expectation = expectationWithDescription(#function)
         let srcIndex = client.getOfflineIndex(#function);
         let dstIndex = client.getOfflineIndex(#function + "_new")
+        srcIndex.beginTransaction()
         srcIndex.addObjects(Array(objects.values)) { (content, error) in
             assert(error == nil)
-            assert(self.client.hasOfflineData(srcIndex.name))
-            assert(!self.client.hasOfflineData(dstIndex.name))
-            self.client.moveIndexOffline(srcIndex.name, to: dstIndex.name) { (content, error) in
-                guard let content = content else { assert(false); return }
-                assert(content["updatedAt"] as? String != nil)
-                assert(!self.client.hasOfflineData(srcIndex.name))
-                assert(self.client.hasOfflineData(dstIndex.name))
-                dstIndex.search(Query(query: "woodstock")) { (content, error) in
+            srcIndex.commitTransaction() { (content, error) in
+                guard error == nil else { assert(false); return }
+                assert(self.client.hasOfflineData(srcIndex.name))
+                assert(!self.client.hasOfflineData(dstIndex.name))
+                self.client.moveIndexOffline(srcIndex.name, to: dstIndex.name) { (content, error) in
                     guard let content = content else { assert(false); return }
-                    guard let nbHits = content["nbHits"] as? Int else { assert(false); return }
-                    assert(nbHits == 1)
-                    expectation.fulfill()
+                    assert(content["updatedAt"] as? String != nil)
+                    assert(!self.client.hasOfflineData(srcIndex.name))
+                    assert(self.client.hasOfflineData(dstIndex.name))
+                    dstIndex.search(Query(query: "woodstock")) { (content, error) in
+                        guard let content = content else { assert(false); return }
+                        guard let nbHits = content["nbHits"] as? Int else { assert(false); return }
+                        assert(nbHits == 1)
+                        expectation.fulfill()
+                    }
                 }
             }
         }
