+++ Tests/OfflineIndexTests.swift
@@ -1,9 +1,24 @@
 //
-//  OfflineIndexTests.swift
-//  AlgoliaSearch
+//  Copyright (c) 2016 Algolia
+//  http://www.algolia.com/
 //
-//  Created by Clément Le Provost on 26/08/16.
-//  Copyright © 2016 Algolia. All rights reserved.
+//  Permission is hereby granted, free of charge, to any person obtaining a copy
+//  of this software and associated documentation files (the "Software"), to deal
+//  in the Software without restriction, including without limitation the rights
+//  to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+//  copies of the Software, and to permit persons to whom the Software is
+//  furnished to do so, subject to the following conditions:
+//
+//  The above copyright notice and this permission notice shall be included in
+//  all copies or substantial portions of the Software.
+//
+//  THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+//  IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+//  FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+//  AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+//  LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+//  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+//  THE SOFTWARE.
 //
 
 import AlgoliaSearch
@@ -41,22 +56,30 @@ class OfflineIndexTests: OfflineTestCase {
     func testAddGetDeleteObject() {
         let expectation = expectationWithDescription(#function)
         let index = client.getOfflineIndex(#function)
+        index.beginTransaction()
         index.addObject(objects["snoopy"]!) { (content, error) in
             guard let content = content else { assert(false); return }
             assert(content["updatedAt"] as? String != nil)
             guard let objectID = content["objectID"] as? String else { assert(false); return }
             assert(objectID == "1")
-            index.getObject("1") { (content, error) in
-                guard let content = content else { assert(false); return }
-                guard let name = content["name"] as? String else { assert(false); return }
-                assert(name == "Snoopy")
-                index.deleteObject("1") { (content, error) in
+            index.commitTransaction() { (content, error) in
+                guard error == nil else { assert(false); return }
+                index.getObject("1") { (content, error) in
                     guard let content = content else { assert(false); return }
-                    assert(content["deletedAt"] as? String != nil)
-                    index.getObject("1") { (content, error) in
-                        assert(error != nil)
-                        assert(error!.code == 404)
-                        expectation.fulfill()
+                    guard let name = content["name"] as? String else { assert(false); return }
+                    assert(name == "Snoopy")
+                    index.beginTransaction()
+                    index.deleteObject("1") { (content, error) in
+                        guard let content = content else { assert(false); return }
+                        assert(content["deletedAt"] as? String != nil)
+                        index.commitTransaction() { (content, error) in
+                            guard error == nil else { assert(false); return }
+                            index.getObject("1") { (content, error) in
+                                assert(error != nil)
+                                assert(error!.code == 404)
+                                expectation.fulfill()
+                            }
+                        }
                     }
                 }
             }
@@ -67,22 +90,30 @@ class OfflineIndexTests: OfflineTestCase {
     func testAddWithIDGetDeleteObject() {
         let expectation = expectationWithDescription(#function)
         let index = client.getOfflineIndex(#function)
+        index.beginTransaction()
         index.addObject(["name": "unknown"], withID: "xxx") { (content, error) in
             guard let content = content else { assert(false); return }
             assert(content["updatedAt"] as? String != nil)
             guard let objectID = content["objectID"] as? String else { assert(false); return }
             assert(objectID == "xxx")
-            index.getObject("xxx") { (content, error) in
-                guard let content = content else { assert(false); return }
-                guard let name = content["name"] as? String else { assert(false); return }
-                assert(name == "unknown")
-                index.deleteObject("xxx") { (content, error) in
+            index.commitTransaction() { (content, error) in
+                guard error == nil else { assert(false); return }
+                index.getObject("xxx") { (content, error) in
                     guard let content = content else { assert(false); return }
-                    assert(content["deletedAt"] as? String != nil)
-                    index.getObject("xxx") { (content, error) in
-                        assert(error != nil)
-                        assert(error!.code == 404)
-                        expectation.fulfill()
+                    guard let name = content["name"] as? String else { assert(false); return }
+                    assert(name == "unknown")
+                    index.beginTransaction()
+                    index.deleteObject("xxx") { (content, error) in
+                        guard let content = content else { assert(false); return }
+                        assert(content["deletedAt"] as? String != nil)
+                        index.commitTransaction() { (content, error) in
+                            guard error == nil else { assert(false); return }
+                            index.getObject("xxx") { (content, error) in
+                                assert(error != nil)
+                                assert(error!.code == 404)
+                                expectation.fulfill()
+                            }
+                        }
                     }
                 }
             }
@@ -93,21 +124,29 @@ class OfflineIndexTests: OfflineTestCase {
     func testAddGetDeleteObjects() {
         let expectation = expectationWithDescription(#function)
         let index = client.getOfflineIndex(#function)
+        index.beginTransaction()
         index.addObjects(Array(objects.values)) { (content, error) in
             guard let content = content else { assert(false); return }
             assert(content["objectIDs"] as? [AnyObject] != nil)
-            index.getObjects(["1", "2"]) { (content, error) in
-                guard let content = content else { assert(false); return }
-                guard let items = content["results"] as? [[String: AnyObject]] else { assert(false); return }
-                assert(items.count == 2)
-                assert(items[0]["name"] as! String == "Snoopy")
-                index.deleteObjects(["1", "2"]) { (content, error) in
+            index.commitTransaction() { (content, error) in
+                guard error == nil else { assert(false); return }
+                index.getObjects(["1", "2"]) { (content, error) in
                     guard let content = content else { assert(false); return }
-                    assert(content["objectIDs"] as? [AnyObject] != nil)
-                    index.getObject("2") { (content, error) in
-                        assert(error != nil)
-                        assert(error!.code == 404)
-                        expectation.fulfill()
+                    guard let items = content["results"] as? [[String: AnyObject]] else { assert(false); return }
+                    assert(items.count == 2)
+                    assert(items[0]["name"] as! String == "Snoopy")
+                    index.beginTransaction()
+                    index.deleteObjects(["1", "2"]) { (content, error) in
+                        guard let content = content else { assert(false); return }
+                        assert(content["objectIDs"] as? [AnyObject] != nil)
+                        index.commitTransaction() { (content, error) in
+                            guard error == nil else { assert(false); return }
+                            index.getObject("2") { (content, error) in
+                                assert(error != nil)
+                                assert(error!.code == 404)
+                                expectation.fulfill()
+                            }
+                        }
                     }
                 }
             }
@@ -118,18 +157,22 @@ class OfflineIndexTests: OfflineTestCase {
     func testSearch() {
         let expectation = expectationWithDescription(#function)
         let index = client.getOfflineIndex(#function)
+        index.beginTransaction()
         index.addObjects(Array(objects.values)) { (content, error) in
             assert(error == nil)
-            let query = Query(query: "snoopy")
-            index.search(query) { (content, error) in
-                guard let content = content else { assert(false); return }
-                guard let nbHits = content["nbHits"] as? Int else { assert(false); return }
-                assert(nbHits == 1)
-                guard let hits = content["hits"] as? [[String: AnyObject]] else { assert(false); return }
-                assert(hits.count == 1)
-                guard let name = hits[0]["name"] as? String else { assert(false); return }
-                assert(name == "Snoopy")
-                expectation.fulfill()
+            index.commitTransaction() { (content, error) in
+                guard error == nil else { assert(false); return }
+                let query = Query(query: "snoopy")
+                index.search(query) { (content, error) in
+                    guard let content = content else { assert(false); return }
+                    guard let nbHits = content["nbHits"] as? Int else { assert(false); return }
+                    assert(nbHits == 1)
+                    guard let hits = content["hits"] as? [[String: AnyObject]] else { assert(false); return }
+                    assert(hits.count == 1)
+                    guard let name = hits[0]["name"] as? String else { assert(false); return }
+                    assert(name == "Snoopy")
+                    expectation.fulfill()
+                }
             }
         }
         waitForExpectationsWithTimeout(expectationTimeout, handler: nil)
@@ -141,13 +184,17 @@ class OfflineIndexTests: OfflineTestCase {
         let settings: [String: AnyObject] = [
             "attributesToIndex": ["foo", "bar"]
         ]
+        index.beginTransaction()
         index.setSettings(settings) { (content, error) in
             assert(error == nil)
-            index.getSettings() { (content, error) in
-                guard let content = content else { assert(false); return }
-                assert(content["attributesToIndex"] as! NSObject == settings["attributesToIndex"] as! NSObject)
-                assert(content["attributesToRetrieve"] as! NSObject == NSNull())
-                expectation.fulfill()
+            index.commitTransaction() { (content, error) in
+                guard error == nil else { assert(false); return }
+                index.getSettings() { (content, error) in
+                    guard let content = content else { assert(false); return }
+                    assert(content["attributesToIndex"] as! NSObject == settings["attributesToIndex"] as! NSObject)
+                    assert(content["attributesToRetrieve"] as! NSObject == NSNull())
+                    expectation.fulfill()
+                }
             }
         }
         waitForExpectationsWithTimeout(expectationTimeout, handler: nil)
@@ -156,16 +203,24 @@ class OfflineIndexTests: OfflineTestCase {
     func testClear() {
         let expectation = expectationWithDescription(#function)
         let index = client.getOfflineIndex(#function)
+        index.beginTransaction()
         index.addObjects(Array(objects.values)) { (content, error) in
             assert(error == nil)
-            index.clearIndex() { (content, error) in
-                guard let content = content else { assert(false); return }
-                assert(content["updatedAt"] as? String != nil)
-                index.browse(Query()) { (content, error) in
+            index.commitTransaction() { (content, error) in
+                guard error == nil else { assert(false); return }
+                index.beginTransaction()
+                index.clearIndex() { (content, error) in
                     guard let content = content else { assert(false); return }
-                    guard let nbHits = content["nbHits"] as? Int else { assert(false); return }
-                    assert(nbHits == 0)
-                    expectation.fulfill()
+                    assert(content["updatedAt"] as? String != nil)
+                    index.commitTransaction() { (content, error) in
+                        guard error == nil else { assert(false); return }
+                        index.browse(Query()) { (content, error) in
+                            guard let content = content else { assert(false); return }
+                            guard let nbHits = content["nbHits"] as? Int else { assert(false); return }
+                            assert(nbHits == 0)
+                            expectation.fulfill()
+                        }
+                    }
                 }
             }
         }
@@ -175,15 +230,19 @@ class OfflineIndexTests: OfflineTestCase {
     func testBrowse() {
         let expectation = expectationWithDescription(#function)
         let index = client.getOfflineIndex(#function)
+        index.beginTransaction()
         index.addObjects(Array(objects.values)) { (content, error) in
             assert(error == nil)
-            index.browse(Query(parameters: ["hitsPerPage": "1"])) { (content, error) in
-                guard let content = content else { assert(false); return }
-                guard let cursor = content["cursor"] as? String else { assert(false); return }
-                index.browseFrom(cursor) { (content, error) in
+            index.commitTransaction() { (content, error) in
+                guard error == nil else { assert(false); return }
+                index.browse(Query(parameters: ["hitsPerPage": "1"])) { (content, error) in
                     guard let content = content else { assert(false); return }
-                    assert(content["cursor"] == nil)
-                    expectation.fulfill()
+                    guard let cursor = content["cursor"] as? String else { assert(false); return }
+                    index.browseFrom(cursor) { (content, error) in
+                        guard let content = content else { assert(false); return }
+                        assert(content["cursor"] == nil)
+                        expectation.fulfill()
+                    }
                 }
             }
         }
@@ -193,16 +252,24 @@ class OfflineIndexTests: OfflineTestCase {
     func testDeleteByQuery() {
         let expectation = expectationWithDescription(#function)
         let index = client.getOfflineIndex(#function)
+        index.beginTransaction()
         index.addObjects(Array(objects.values)) { (content, error) in
             assert(error == nil)
-            index.deleteByQuery(Query(parameters: ["numericFilters": "born < 1970"])) { (content, error) in
-                assert(error == nil)
-                index.browse(Query()) { (content, error) in
-                    guard let content = content else { assert(false); return }
-                    guard let nbHits = content["nbHits"] as? Int else { assert(false); return }
-                    assert(nbHits == 1)
-                    assert(content["cursor"] == nil)
-                    expectation.fulfill()
+            index.commitTransaction() { (content, error) in
+                guard error == nil else { assert(false); return }
+                index.beginTransaction()
+                index.deleteByQuery(Query(parameters: ["numericFilters": "born < 1970"])) { (content, error) in
+                    assert(error == nil)
+                    index.commitTransaction() { (content, error) in
+                        guard error == nil else { assert(false); return }
+                        index.browse(Query()) { (content, error) in
+                            guard let content = content else { assert(false); return }
+                            guard let nbHits = content["nbHits"] as? Int else { assert(false); return }
+                            assert(nbHits == 1)
+                            assert(content["cursor"] == nil)
+                            expectation.fulfill()
+                        }
+                    }
                 }
             }
         }
@@ -212,17 +279,21 @@ class OfflineIndexTests: OfflineTestCase {
     func testMultipleQueries() {
         let expectation = expectationWithDescription(#function)
         let index = client.getOfflineIndex(#function)
+        index.beginTransaction()
         index.addObjects(Array(objects.values)) { (content, error) in
             assert(error == nil)
-            index.multipleQueries([Query(query: "snoopy"), Query(query: "woodstock")]) { (content, error) in
-                guard let content = content else { assert(false); return }
-                guard let results = content["results"] as? [[String: AnyObject]] else { assert(false); return }
-                assert(results.count == 2)
-                for result in results {
-                    guard let nbHits = result["nbHits"] as? Int else { assert(false); return }
-                    assert(nbHits == 1)
+            index.commitTransaction() { (content, error) in
+                guard error == nil else { assert(false); return }
+                index.multipleQueries([Query(query: "snoopy"), Query(query: "woodstock")]) { (content, error) in
+                    guard let content = content else { assert(false); return }
+                    guard let results = content["results"] as? [[String: AnyObject]] else { assert(false); return }
+                    assert(results.count == 2)
+                    for result in results {
+                        guard let nbHits = result["nbHits"] as? Int else { assert(false); return }
+                        assert(nbHits == 1)
+                    }
+                    expectation.fulfill()
                 }
-                expectation.fulfill()
             }
         }
         waitForExpectationsWithTimeout(expectationTimeout, handler: nil)
